feat(domains): unify single and bulk add into one dialog (#495)
