+++ app/features/edge/domain/add/index.ts
@@ -1,0 +1,1 @@
+export { AddDomainsDialog } from './add-domains-dialog';
