--- app/features/edge/domain/bulk-add/index.ts
@@ -1,1 +1,0 @@
-export { BulkAddDomainsAction } from './bulk-add-domains-action';
