+++ app/routes/project/detail/domains/index.tsx
@@ -2,11 +2,7 @@ import { BadgeCopy } from '@/components/badge/badge-copy';
 import { useConfirmationDialog } from '@/components/confirmation-dialog/confirmation-dialog.provider';
 import { NameserverChips } from '@/components/nameserver-chips';
 import { createActionsColumn, Table } from '@/components/table';
-import { BulkAddDomainsAction } from '@/features/edge/domain/bulk-add';
-import {
-  DomainFormDialog,
-  type DomainFormDialogRef,
-} from '@/features/edge/domain/domain-form-dialog';
+import { AddDomainsDialog } from '@/features/edge/domain/add';
 import { DomainExpiration } from '@/features/edge/domain/expiration';
 import { DomainStatus } from '@/features/edge/domain/status';
 import { PermissionButton, useResourcePermissions } from '@/modules/rbac';
@@ -39,8 +35,8 @@ import { toast } from '@datum-cloud/datum-ui/toast';
 import { Tooltip } from '@datum-cloud/datum-ui/tooltip';
 import { useQueryClient } from '@tanstack/react-query';
 import { ColumnDef } from '@tanstack/react-table';
-import { GlobeIcon, ListChecksIcon, PlusIcon, TrashIcon } from 'lucide-react';
-import { useMemo, useCallback, useEffect, useRef, useState } from 'react';
+import { GlobeIcon, PlusIcon, TrashIcon } from 'lucide-react';
+import { useMemo, useCallback, useEffect, useState } from 'react';
 import { LoaderFunctionArgs, useNavigate, useParams, useSearchParams } from 'react-router';
 
 type FormattedDomain = {
@@ -182,14 +178,13 @@ function DomainsInner({
   const { confirm } = useConfirmationDialog();
   const { enqueue, showSummary } = useTaskQueue();
   const { project, organization } = useApp();
-  const domainFormRef = useRef<DomainFormDialogRef>(null);
-  const [bulkAddPopoverOpen, setBulkAddPopoverOpen] = useState(false);
+  const [addOpen, setAddOpen] = useState(false);
 
   // Open create dialog from URL search params (e.g. ?action=create)
   useEffect(() => {
     if (searchParams.get('action') === 'create') {
       if (canCreate) {
-        domainFormRef.current?.show();
+        setAddOpen(true);
       }
       const nextParams = new URLSearchParams(searchParams);
       nextParams.delete('action');
@@ -533,23 +528,24 @@ function DomainsInner({
         description="Manage domains as programmatic resources no matter where they are registered, or where the DNS is hosted. Note: verification of domain ownership is required for some features."
         search="Search"
         actions={[
-          canCreate ? <BulkAddDomainsAction key="bulk-add" projectId={projectId!} /> : null,
-          <PermissionButton
-            key="create"
-            resource="domains"
-            verb="create"
-            group="networking.datumapis.com"
-            scope="project"
-            deniedReason="You don't have permission to add a domain"
-            type="primary"
-            theme="solid"
-            size="small"
-            className="w-full sm:w-auto"
-            data-e2e="create-domain-button"
-            onClick={() => domainFormRef.current?.show()}>
-            <Icon icon={PlusIcon} className="size-4" />
-            Add domain
-          </PermissionButton>,
+          canCreate ? (
+            <PermissionButton
+              key="add"
+              resource="domains"
+              verb="create"
+              group="networking.datumapis.com"
+              scope="project"
+              deniedReason="You don't have permission to add a domain"
+              type="primary"
+              theme="solid"
+              size="small"
+              className="w-full sm:w-auto"
+              data-e2e="create-domain-button"
+              onClick={() => setAddOpen(true)}>
+              <Icon icon={PlusIcon} className="size-4" />
+              Add domains
+            </PermissionButton>
+          ) : null,
         ].filter(Boolean)}
         multiActions={
           canDelete
@@ -573,41 +569,28 @@ function DomainsInner({
           actions: canCreate
             ? [
                 {
-                  label: 'Add domain',
+                  label: 'Add domains',
                   type: 'button',
                   icon: <Icon icon={PlusIcon} className="size-3" />,
-                  onClick: () => domainFormRef.current?.show(),
-                },
-                {
-                  label: 'Bulk add domains',
-                  type: 'button',
-                  variant: 'outline',
-                  icon: <Icon icon={ListChecksIcon} className="size-3" />,
-                  onClick: () => setBulkAddPopoverOpen(true),
+                  onClick: () => setAddOpen(true),
                 },
               ]
             : [],
         }}
       />
 
-      <DomainFormDialog
-        ref={domainFormRef}
+      <AddDomainsDialog
         projectId={projectId!}
-        onSuccess={(domain) => {
+        open={addOpen}
+        onOpenChange={setAddOpen}
+        onSuccess={(domain) =>
           navigate(
             getPathWithParams(paths.project.detail.domains.detail.overview, {
               projectId,
               domainId: domain.name,
             })
-          );
-        }}
-      />
-
-      {/* Controlled BulkAddDomainsAction for empty content button - renders as Dialog */}
-      <BulkAddDomainsAction
-        projectId={projectId!}
-        popoverOpen={bulkAddPopoverOpen}
-        onPopoverOpenChange={setBulkAddPopoverOpen}
+          )
+        }
       />
     </>
   );
